+++ source/App/documents/release-notes/release-notes.md
@@ -1,5 +1,13 @@
 # App Release notes
 
+## Version 12.1.0
+
+- In `FunctionApp` project:
+    - Add `Middleware.UserMiddleware`
+    - Add reusable dependency injection extensions:
+        - Add `AuthenticationBuilderExtensions.UseUserMiddlewareForIsolatedWorker<TUser>`
+        - Add `AuthenticationExtensions.AddUserAuthenticationForIsolatedWorker<TUser, TUserProvider>`
+
 ## Version 12.0.1
 
 - No functional change